git-svn-id: svn://fileserver/activex/AVS/Sources/TeamlabOffice/trunk/ServerComponents@52610 954022d7-b5bf-4e40-9824-e11837661b57

+++ Common/DocxFormat/Source/XlsxFormat/Worksheets/Worksheet.h
@@ -151,7 +151,7 @@ namespace OOX
 								pCommentItem->m_nCol = nCol - 1;
 
 								unsigned int nAuthorId = pComment->m_oAuthorId->GetValue();
-								if(nAuthorId < aAuthors.GetSize())
+								if(nAuthorId < (unsigned int)aAuthors.GetSize())
 									pCommentItem->m_sAuthor = *aAuthors[nAuthorId];
 
 								OOX::Spreadsheet::CSi* pSi = pComment->m_oText.GetPointerEmptyNullable();
